pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sam3d"
version = "0.1.0"
description = "Unified Python interface for Meta SAM 3D Body and SAM 3D Objects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.optional-dependencies]
inference = []

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
